refactor: separate create user dialog from create user form
